refactor(enums): add subfolders for enums
