+++ app/Enums/Lessons/LessonOrder.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Enums;
+namespace App\Enums\Lessons;
 
 enum LessonOrder: int
 {
