+++ app/Enums/Shared/DayOfWeek.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Enums;
+namespace App\Enums\Shared;
 
 use Filament\Support\Contracts\HasLabel;
