+++ app/Models/Exception.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Models;
 
-use App\Enums\LessonOrder;
+use App\Enums\Lessons\LessonOrder;
 use Database\Factories\ExceptionFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
