+++ app/ValueObjects/LessonValueObject.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\ValueObjects;
 
-use App\Enums\LessonOrder;
+use App\Enums\Lessons\LessonOrder;
 use Carbon\CarbonImmutable;
 use Closure;
